// madar-backend/src/common/enums/user-role.enum.ts
export enum UserRole {
  STUDENT = 'student',
  COMPANY = 'company',
  UNIVERSITY = 'university',
  ADMIN = 'admin',
  SUPER_ADMIN = 'super_admin',
  COORDINATOR = 'coordinator',
  UNIVERSITY_VIEWER = 'university_viewer',
  DATA_OFFICER = 'data_officer',
  QUALITY_OFFICER = 'quality_officer',
  ACADEMIC_DEVELOPMENT_OFFICER = 'academic_development_officer',
}